backend/requirements.txt
# Flask Core
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0

# Database
pymongo==4.6.1
dnspython==2.4.2

# Security & Validation
bcrypt==4.1.2
marshmallow==3.20.1

# Environment & Configuration
python-dotenv==1.0.0

# AI Integration
google-generativeai==0.3.2

# Utilities
python-dateutil==2.8.2
requests==2.31.0

# Development (optional)
gunicorn==21.2.0